Add files via upload
HEY So I believe I have created a list for the bullets and the enemy's and inserted a pointer to them into the list. I am struggling trying to get them deleted when an enemy collides with a bullet. I've tried declaring the lists globally, that doesn't work. Not sure if we hve to restructure the program or what I am doing wrong but I am trying. Hector/Freddie do you have time to look at this? I am at a wall and have to leave for work at 1130.