Add files via upload
I put in changes to how the levels get harder. For now, I made it so that it increase one enemy per everytime the scene is cleared of enemies. Tomorrow imma trouble shoot the bullets error with jamison. The mainwindow file I'm adding may not have what you guys may have added, so my bad if I dont have that in there.