Add files via upload
I changed player so that it would give x/y values to the bullet to start in a somewhat close place to shoot. I also set the game to reset to level 1 if the player presses 'new'. I haven't accounted for the ending on what should show, etc. I will work with eric on this.